Moved some helper functions to globalhelper.
