+++ tests/lifecycle/lifehelper/lifehelper.go
@@ -62,31 +62,6 @@ func DefinePod(name string) *corev1.Pod {
 		globalhelper.Configuration.General.TestImage)
 }
 
-// CreateAndWaitUntilStatefulSetIsReady creates statefulset and wait until all replicas are ready.
-func CreateAndWaitUntilStatefulSetIsReady(statefulSet *v1.StatefulSet, timeout time.Duration) error {
-	runningReplica, err := globalhelper.APIClient.StatefulSets(statefulSet.Namespace).Create(
-		context.Background(),
-		statefulSet,
-		metav1.CreateOptions{})
-	if err != nil {
-		return err
-	}
-
-	Eventually(func() bool {
-		status, err := isStatefulSetReady(runningReplica.Namespace, runningReplica.Name)
-		if err != nil {
-			glog.V(5).Info(fmt.Sprintf(
-				"statefulSet %s is not ready, retry in %d seconds", runningReplica.Name, lifeparameters.RetryInterval))
-
-			return false
-		}
-
-		return status
-	}, timeout, lifeparameters.RetryInterval*time.Second).Should(Equal(true), "statefulSet is not ready")
-
-	return nil
-}
-
 // CreateAndWaitUntilReplicaSetIsReady creates replicaSet and wait until all replicas are ready.
 func CreateAndWaitUntilReplicaSetIsReady(replicaSet *v1.ReplicaSet, timeout time.Duration) error {
 	runningReplica, err := globalhelper.APIClient.ReplicaSets(replicaSet.Namespace).Create(
@@ -108,32 +83,6 @@ func CreateAndWaitUntilReplicaSetIsReady(replicaSet *v1.ReplicaSet, timeout time
 
 		return status
 	}, timeout, lifeparameters.RetryInterval*time.Second).Should(Equal(true), "replicaSet is not ready")
-
-	return nil
-}
-
-// CreateAndWaitUntilPodIsReady create and wait until pod is in a "Running" phase.
-func CreateAndWaitUntilPodIsReady(pod *corev1.Pod, timeout time.Duration) error {
-	createdPod, err := globalhelper.APIClient.Pods(pod.Namespace).Create(
-		context.Background(),
-		pod,
-		metav1.CreateOptions{})
-	if err != nil {
-		return err
-	}
-
-	Eventually(func() bool {
-		status, err := isPodReady(createdPod.Namespace, createdPod.Name)
-		if err != nil {
-
-			glog.V(5).Info(fmt.Sprintf(
-				"deployment %s is not ready, retry in %d seconds", createdPod.Name, lifeparameters.RetryInterval))
-
-			return false
-		}
-
-		return status
-	}, timeout, lifeparameters.RetryInterval*time.Second).Should(Equal(true), "Deployment is not ready")
 
 	return nil
 }
@@ -174,25 +123,6 @@ func WaitUntilClusterIsStable() error {
 	return err
 }
 
-func isStatefulSetReady(namespace string, statefulSetName string) (bool, error) {
-	testStatefulSet, err := globalhelper.APIClient.StatefulSets(namespace).Get(
-		context.Background(),
-		statefulSetName,
-		metav1.GetOptions{},
-	)
-	if err != nil {
-		return false, err
-	}
-
-	if testStatefulSet.Status.ReadyReplicas > 0 {
-		if testStatefulSet.Status.Replicas == testStatefulSet.Status.ReadyReplicas {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
 func isReplicaSetReady(namespace string, replicaSetName string) (bool, error) {
 	testReplicaSet, err := globalhelper.APIClient.ReplicaSets(namespace).Get(
 		context.Background(),
@@ -207,24 +137,6 @@ func isReplicaSetReady(namespace string, replicaSetName string) (bool, error) {
 		if testReplicaSet.Status.Replicas == testReplicaSet.Status.ReadyReplicas {
 			return true, nil
 		}
-	}
-
-	return false, nil
-}
-
-func isPodReady(namespace string, podName string) (bool, error) {
-	podObject, err := globalhelper.APIClient.Pods(namespace).Get(
-		context.Background(),
-		podName,
-		metav1.GetOptions{},
-	)
-
-	if err != nil {
-		return false, err
-	}
-
-	if podObject.Status.Phase == "Running" {
-		return true, nil
 	}
 
 	return false, nil
